Consolidate all tests into single test class
